[Graphics]
Removed VertexCount variable from the Mesh class and replaced it with a define since it shouldn't change during the course of the program.
Changed how the Graphics system renders quads. The rendering system is now batched so that all the quads should be in one draw call.
Overloaded the VAOPrepare function to accept vectors of mesh data rather than individual graphics components.
Added a new function BatchPrepare to gather all the mesh data from the given entities for the rendering.