[Graphics][Serialization][Engine]
    ENTITY .h/.cpp
Changed the constructors, the default constructor will init the component to nullptr, the non default constructor takes a transform and graphics components.
Added setter for the graphics and transform components.
Make the getters const.

    GRAPHICS COMPONET .h/.cpp
Added a copy constructor for the Mesh class.
Compressed vertex data into matrices rather than vectors.
Added a function to set the color of a mesh.
Expanded the Texture class.
Added an enum that describes the texture constructed
Added a Save and Load function for serialization.
Expanded the Shader class.
Added an enum that describes the shader constructed.
Changed the constructor to take file paths rather than a GLuint.
Added a Save and Load function for serialization.
Changed the graphics component constructor to take a texture type and a shader type.
Added a copy construtor for the graphics component.
Added a texture type and shder type to the graphics component.
Moved all the implementation for the functions to the .cpp file.

    GRAPHICS SYSTEM .h/.cpp
Added a texture manager.
Added a variable to keep track of the number of entities each update.
Added a check for the number entities in the update that will prevent unnecessarily updating the batch rendering data and VAO.
Added textures to the rendering.
Changed how the camera moves based on zoom level, it will move faster or slower if you are zoomed out or in respectively.
Added a Texture UV VBO to the VAO.
Reduced the number of matrix multiplications in the BatchPrepare function from 4 to 1 by having the position vectors composited into a matrix.

    SERIALIZER .h
Created a serialization class that is constructed with a file path to the json file and an enum mode, read or write.
Created a templated Read and Write function that will call the given object's Load or Save function respectively

    Shader Manager .h/.cpp
Changed the GetShaderID function to take a ShaderType.
Changed the BuildShader function to take a  reference to a shader object.
The init function not constructs a serializer object to read shaders from a json file and build them one at a time.

    TEXTURE MANAGER .h/.cpp
Created a texture manager that is very similar to the shader manager. 

    TRANSFORM COMPONENT .h/.cpp
Added a copy constructor.

    ENGINE.cpp
Changed the entities were being initialized in the constructor. The number of entities being made is easier to read, they start from position (0, 0) now, the randomness of the  additional color is reduced, and now they are made to look like a flat plane in the isometric perspective. 
The shutdown function now destroys all the entities created in the constructor.