[Mesh]
    MESH MANAGER .h/.cpp
Added a mesh manager class. This will reduce the memory usage.
    BASE SYSTEM .h
Added the skeleton of a class for standard system calls.
    GRAPHICS COMPONENT .h/.cpp
Removed the Mesh pointer from the Graphics Component and replaced it with a glm::vec4 for the color.
Removed the color variable from Mesh.
Commented much of the file.
    GRAPHICS SYSTEM .h/.cpp
Added a Mesh Manager.
Changed a couple functions to use the new Mesh Manager rather than getting the mesh from the Graphics Component.
    TRANSFORM COMPONENT .h/.cpp
Changed the scale to be a single float rather than a vec2. All entities are uniformly scaled and this will save a bit of memory.
Changed the position from a vec3 to a vec2. We are only rendering in a 2D space and this will save a bit of memory.